Exported articles from Default Layout component as separate components
